Create ExampleOptimizer in separate header
